Added in-game UI display for hp, ammo and score

+++ player.h
@@ -11,6 +11,6 @@ extern player_t player;
 void player_init();
 int player_menu();
 int player_update();
-void player_hit(int damage);
+void player_hit(uint8_t damage);
 
 #endif
